Changes to support mode switching
Changes in the RuntimeMetadata ecore:
- modeChart attribute at ComponentInstance ecore class
- mode attribute at ComponentProcess ecore class
All other changes in the RuntimeMetadata-generated files are not
important - they are just because of re-generating the files with the
latest ecoregen implementation (which moves some "</p>" up in the
generated comments).

Also, added DEECoMode interface and ModeChart abstract class to be
implemented/extended by application-specific classes related to
mode-switching.